requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
psycopg2-binary
requests
pydantic
openai
redis
slowapi
structlog
bleach
python-multipart
thefuzz
python-levenshtein
aiofiles
httpx
passlib[bcrypt]
python-jose[cryptography]
